Initial commit
This gym simulates a very basic cloud provider environment with a service that processes requests from a queue.
The environment simulates the queue, instances with a certain processing capacity and latency for instance start up.
The reward function gives feedback about how well the boundaries are maintained, how well the queue size is kept minimal
and how well the cloud provider's resources are utilized related to the incoming load.